cv_service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
Pillow>=10.0.0
numpy>=1.24.0
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.0
transformers>=4.37.0
scikit-image>=0.22.0
scipy>=1.11.0
python-multipart>=0.0.9
